+++ src/Network/Network/Models/PSFlowLogResource.cs
@@ -35,6 +35,9 @@ namespace Microsoft.Azure.Commands.Network.Models
         public string EnabledFilteringCriteria { get; set; }
 
         [Ps1Xml(Target = ViewControl.Table)]
+        public string RecordType { get; set; }
+
+        [Ps1Xml(Target = ViewControl.Table)]
         public bool? Enabled { get; set; }
 
         [Ps1Xml(Target = ViewControl.Table)]